Clean things up to be a lot more general.
In particular, we now go ahead and parse *all* package.xml
files we can find at the beginning.  This ends up doing doing
a bunch of extra work (we'll parse package.xml files that we
don't need), but it will be much more correct than just depending
on the path name.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>